Move automatic creation of namespace inside of Team observer
Previously the creation of a new user triggered the creation of a
team with the same user (owned by the user) and of a namespace with
the same name of the user.
Creating a new team resulted in no namespace being created.

This commit introduced a new observer for the Team model. Every time
a user is created a new team is being created, exactly like before.
However the namespace creation is delegated to the team observer.
In that way creating a team results in creating also an associated
namespace.