Fix flaicky test
It turns out factory_girl's `attributes_for` returns always the same
instances for the attributes. Some tests changed these attributes to
reproduce corner cases, leading to altering the values used by the
other tests. The failure never presented because rspec executes the
tests in a random order.

This patch avoids the usage of `attributes_for` and ensures the
tests using raw_events operate on different instances of the objects.