Introducing the special 'portus' user
Soon we are going to fetch informations from the registry itself (like
repositories' manifests). When doing that we have to authenticate
against the registry like all normal clients. Basically Portus is going to
authenticate against itself. :)

The client library used to interact with the registry is already merged
in master (see `app/models/registry_client.rb`). However we still have
to provide a username and password. Hence we need to send the password
as a client, and then verify itself inside of the token controller.

We are going to authenticate as 'portus' user providing a time based one
time only password (aka TOTP).
The TOTP is generated using a secret that changes every time the Rails process
starts (see `config/application.rb`).

The policy checks inside of the token controller have been disabled for the
'portus' user because he can access to everything.s

I've also reserved 'portus' as username to avoid misuses from the users.