registrations: don't show "Create Admin" on signup error if there is an admin
Devise does not redirect to the 'new' page on signup error, but it renders it.
This means that on signup error the `@admin` view variable will be nil, thus
showing the "Create Admin" button. In this commit a before_filter has been
added for both the 'new' and the 'create' actions, so this case is always
handled consistently.

Signed-off-by: Miquel Sabaté <mikisabate@gmail.com>