[CI] run pending migrations in production when initializing
this way we don't need to run "rake db:migrate" on a post install step
in the RPM, but rails does that when restarting apache, which will
happen when installing/updating portus in any case, whether we use the
RPM to deploy or not.

Skip migrations when precompiling assets in the package.