Ported flashy messages from #notice to #alert
A ton of JS expected the #notice element to exist but that was never the case.
Instead, the element to target for flashy messages is the #alert one. This must
be some leftover from a refactoring.

Moreover, I've added some acceptance tests that were missing (e.g. Team
members), since some of the alert messages referenced this features anyways.