Fixed the Portus user for the RegisterClient class
Up until now, the portus "user" was handled in a special way, by adding some
code that worked around the whole authentication mechanism. However, we
recently discovered that this was basically broken.

The solution to this situation is to effectively create the "portus" user, as
any other user in the system. Because of this, I've added a new secret, which
is the password of the "portus" user, and I have added a rake task to help with
the migration. Moreover, the seeds file already takes care of that. With this
commit, the `manifest` method from the RegisterClient works again, and the
authentication mechanism has been greatly simplified.

There are still some small issues that need to be addressed. This will be done
in later commits.