Fix the creation of admins and the listing of users after introducing portus
The portus user has to be handled transparently behind the scenes. This means
that it must not be listed in the users list for admins, and it has to be
hidden when dealing with authentication.

Moreover, the admin will have a notice in the dashboard if the Portus user is
not there. This should never happen since the portus user is installed always
by seeding it. This would mean that there's something really wrong with the DB
(some manual update gone wrong).