Take advantage of the Catalog API
In recent commits, the RegistryClient not only was fixed, but it also
introduced the catalog method. With this commit we take advantage of this new
method by calling it in a new crono job.

For managing and launching this new job we use the crono gem as suggested by
@lsamayoa.

Fixes #199