Implement error controller for requirements and for database exceptions
Implement an error controller to show custom messages when requirements
like ssl configuration or secrets configurations are not met. Also, set
up this controller as a custom error handler for showing up a custom
message if database is not accessible.

do not exit when running database migrations on production but
instead show the error page since any access to the database will
raise an exception.

This way we have a running application that states the configuration
errors.