api/v2: fixed pull permissions for private namespaces and unauthorized users
The commit 05eb390776ee699e74947ccf19b2d58e642f22c5 introduces a regression in
which a logged in user can pull a repo from a private namespace that doesn't
belong to it.

The problem is that the delete_if method is screwing with the `each` iteration.
To avoid this, the returned scopes from `scope_handler` is a duplicate, instead
of being the same object.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>