Allow the admin to enable disabled users
In order to do this, the previous `disable` method has been transformed into
the `toggle_enabled` method. This method transparently toggles the enabled
column accordingly. This has involved quite some changes on the code (plus, it
has been moved to the model, instead of keeping the controller fat).

On the admin page, both enabled/disabled users are displayed, so the admin has
the ultimate power to enable users back.

Fixes #245

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>