Respect the "gravatar" configurable value everywhere
I realized that in the public activities we use `gravatar_image_tag` directly,
without caring whether the current Portus instance has gravatar support enabled
or not. Therefore, I'm introducing a helper method that abstracts this away,
and I've replaced all the wrong uses of the `gravatar_image_tag` with this new
helper method.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>