registry: fixed webhooks for distribution v2.1
Distribution changed the format of event notifications. In particular, the tag
is no longer provided inside the URL. This is a good move since it's safer to
point by digest. However it affected us because we retrieved tags from this
URL. From now on, we'll take advantage of this digest to call the Manifest API
and fetch the tag.

I've also submitted an issue on Docker Distribution about providing a better
solution for this (so we don't have to call the Manifest API everytime).

See docker/distribution#966

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>