Simplified the RegistryNotification and the RegistryClient classes
The notifications system consisted of an abstraction of three classes. Those
classes were a bit magical and weren't documented. My approach has been to
simplify all of this and implementing it all in one library class. This makes
sense because all the notifications that we catch are `repository` related
anyways.

The RegistryClient will pick the "portus" user by default, which is what we
want anyways usually.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>