[packaging] review registry and crono scripts
Rename configure_crono to configure_crono.sh and set
executable mode.

Add the registry.config.yml.in file which was missing.

Enable delete feature in registry by default.

Use environment variables in the scripts. When running rake
in the scripts, it was failing because we didn't have the
environment variables set. Let's move the rake command to where we
set those variables. This way we can run "rake db:seed"

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>