ldap: added support for SSL/TLS connections
As expected, the Net::LDAP class supports this seamlessly. Moreover, the added
configuration is minimal: there is now the "method" configurable value with
three possible values ("plain", "simple_tls" and "starttls").

I'd say that "plain" is the best default, since it's the easiest to configure
from a server point of view. However, the recommended choice is "starttls",
as described in the documentation of this newly added value.

I've also relaxed a rubocop rule. I plan to fix this as soon as possible.

Fixes #311

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>