Added SSL support for the communication between Portus and the registry
This fixes some known issues like the setup of the appliance, in which the
Portus instance and the registry just couldn't speak to each other.

In order to do this, the `use_ssl` attribute has been introduced. This
attribute can be changed in the "registries#show" page in the form of a toggle
box. There's also a checkbox that allows the admin to set `use_ssl` as true
when creating a new registry.

Moreover, this commit also brings some needed cleaning in regards to classes
dealing with registries.

Fixes #362

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>