registry: return the proper error when fetching the manifest fails
As noted by @morodin in the issue #338, when Portus fails the fetch the
manifest, it adds a new tag "1" into the database. This is due to a regression
introduced in 09dc052721ad560e4fb56c585f6193d3305fc578. In this case, on
failure the `get_tag_from_manifest` method returned true on error, because
that's what the logger returns. This commit makes sure that in this case nil is
returned.

Moreover, Portus will now also log the error that has happenned when fetching
the manifest.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>