catalog_job: skip repos belonging to a non-existing namespace
When fetching the catalog, if a given repo is supposed to belong to a namespace
that doesn't exist, do nothing. This is more likely to be a problem in the
registry, since we do not allow to create repos in namespaces that are not in
the system.

This commit also closes the possibility for the CatalogJob to be used as an
automatic migration tool.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>