registries: allow the admin to edit a registry
An admin will be able from now on to edit both the `name` and the `use_ssl`
attributes from a registry. The `hostname` attribute will only be available for
update if there are no repositories in the system yet. This is done this way to
prevent the admin from performing unsafe actions, while allowing some
flexibility when setting up Portus for the first time.

See #408

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>