Make sure that the registry is reachable when creating it
When creating a new registry, Portus will now check whether it's reachable or
not. If it is, it will continue as usual. However, if it's not reachable, then
Portus will redirect to the registries#new page and alert the user with some
useful message. When this redirect happens, the user will have available an
extra checkbox called "Skip remote checks". If the user checks it, then it's
telling to Portus to force the creation of the registry, even if it's not
reachable.

Fixes #408

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>