Allow controlled anonymous access with public namespaces
Before this patch, a "public" namespace allowed pulls to any Portus user.
However, it did not allow pull access to anonymous users.

After this patch, the TokensController won't authenticate every user, but only
those that have called the `docker login` command. From now on, it's up to
each
policy to allow/disallow anonymous users. In particular, the only one that
allows access to anonymous users is the Namespace policy, and it only does it
for pull requests.

Moreover, the UI toggles have now a title that tells the user that by making a
namespace public, anyone will be able to pull.

Fixes #279

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>