compose: do not allow the pwd to be another than "portus"
This is known to be problematic, since our docker compose setup expects the
portus_web container to be created. The problem that this is based on the pwd
alone. It's not possible to specify a name for this, as reported on the
following issue: https://github.com/docker/compose/issues/2092.

Thanks to @jjperezaguinaga for investigating this.

Fixes #469

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>