public_activity: simplify messages of users acting upon themselves
Users can remove themselves from a team, and owners can step down as owners of
a team. In these cases, it's weird to see a message like "mssola removed user
mssola ...". Instead, now the same message will say "mssola removed itself
...".

Moreover, in the case of removing a team member, it's weird in my opinion to
state the old role of that user. So, comparing on the removal of a team member:

- before: "mssola removed user mssola with role owner from the docker team"
- after:  "mssola removed itself from the docker team"

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>