passwords: do not allow LDAP users to access the password reset procedure
It doesn't make sense for LDAP users to recover passwords this way, since
Portus acts just as a proxy between users and the LDAP server when it comes to
authentication.

Since this was exactly the same as for the signup page, I've put the code
inside of a concern that bith the RegistrationsController and the
PasswordsController use.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>