Added the `signup` configurable value
When enabled (default behavior), then any user can access to the signup page.
Otherwise, users are not able to enter the signup form. This is ignored in LDAP
mode.

Note that I've also added the `verify` configurable value. This is supposed to
be implemented in the near future, as described in the related issues.

See issues #179 and #283

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>