compose: do nothing if the DOCKER_HOST variable is already set
This fixes the setup on environments that already set the DOCKER_HOST
environment variable. Since this variable can be set to something like
"tcp://IP:POR", the final message cleans up everything so only the IP is shown
together with the proper ports.

I've also cleaned up the final message so it's more well-formatted and verbose.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>