Fixed the crono job overflowing the `log` column
Before this commit, the crono job kept on adding the log into the `log` column
from the `crono_jobs` table. This is, of course, a bad thing. What I did to fix
this is to simply remove this columns from the DB, and crono is ok with that.
Running crono again with this patch runs in the very same way as it used to be,
but without adding the log all the time.

Moreover, the log is supposed to go to the stdout, as envisioned by the
Twelve-Factor App manifesto: http://12factor.net. Docker follows this too, and
by doing so then the log of this job blends in with docker-compose logs
command. This has two main consequences:

1. The user can decide how to manage logs in an easier way, so there's no way
   that we run out of disk (an issue we were having).
2. supervisor is no longer needed. Instead, we take advantage of the
   compose-setup.sh script to do it properly. Before this patch, for some
   reason the catalog job did not reach the db container.

Fixes #479
Fixes #505

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>