lib/portus: do not remove all the repos if one fails
Right now, the `RegistryClient#catalog` method can erase the DB of all the
repos. This happens when, for some unexpected reason, the
`RegistryClient#add_tags` method fails at retrieving one repository. In this
case, before this patch this method just returned an empty array. After this
patch, repositories that are not found will simply not be added, but the
method will go on adding tags to other repositories.

See #663

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>