app: always use _path methods instead of *_url
The _url methods should only be used when the full path (+ hostname, etc.) is
needed. This is usually only the case for things like emails and views that are
to be accessed from outside the web app. This can usually be workarounded
through NGinx and similar tools, but it's better to have this behavior even
without such tools.

See #677

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>