handle multiple scopes in token requests
According to the official spec the scope parameter can be passed more than
once: "This query parameter should be specified multiple times ifthere is more
than one scope entry from the WWW-Authenticate header".
At least the latest docker daemon 1.10.0 does use multiple scope parameters in
a single token request (e.g. when pushing multiple tags in parallel).

Additionaly this commit changes the token endpoint to not return
"401 Unauthorized" on any kind of authorization error, only authentication
errors should return 401 errors.
This also conforms with the spec: "If the client only has a subset of the
requested access it must not be considered an error as it is not the
responsibility of the token server to indicate authorization errors as part of
this workflow."

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>