Don't allow usernames to clash with teams
Up until now we allowed users to have the same username as some teams. This is
troublesome because then the private team assigned to this user would clash
with the existing one.

In the future we should figure out a way to better handle these hidden teams
(e.g. by creating one with a random name and the reference it on the user row).
For now, in order to not be too disruptive we will disallow creating users with
the same name as an existing team (note that the other way around is also
disallowed).

Fixes #836

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>