spec/integration: added integration testing
This commit is the foundation for integration testing in the Portus repository.
It uses docker containers to setup everything, and a lot of magic has been
written around RSpec so it integrates with the style of the other tests.

This commit does not handle on purpose the following (which have been submitted
as separate issues):

  - Test multiple versions of Docker.
  - Test crono
  - Test portusctl
  - Cover as much of the registry client as possible
  - Add integration testing into CI

Fixes #745

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>