Add the Image ID in the repositories#show page
This is the same Image ID that is shown in the docker client. In order to do
this, a new column has been added to the `tags` table. The registry client has
also been modified to handle all this more gracefully. The image ID being
shown is truncated in the exact same format as in the Docker client. That
being
said, users can hover the mouse over the image ID and they'll get the full
image ID.

Fixes #512

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>