portusctl: disable automatic generation of certificates
Introduce two new flags regarding SSL certificates:

	* --ssl-gen-self-signed-certs
	* --ssl-certs-dir <dir>

The former, as the name suggests, creates self-signed certificates. The
latter uses certificates located in the specified directory. From there,
it will copy the certificates to the apache directory and continue the
process as before.
Should either one of the certificates be missing, it will fail with a
meaningful message.

If neither flag is specified, it assumes the certificates are already
located in `/etc/apache2/ssl.{key,crt}/`.
As before, should either one of the certificates be missing, it will
fail with a meaningful message.

These two flags are mutually exclusive.

Signed-off-by: Thomas Hipp <thipp@suse.com>