Allow the FQDN to be specified through config
This is something in between to what we have on the `v2.0` branch and the
`master` branch. This way we have the behavior from `master` while not breaking
what we had in previous patch-level releases.

This commit is largely based on 984671662ade8c105f019b1ee108b71ef87a9821 in the
`master` branch.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>