Review release script: tarball name and version of RPM
Because when releasing the service is disabled, the tarball name
has changed and we need to fix that. Also, we need to remove
the previous tarball.

Remove date from version in rpm and instead, have the version in VERSION file,
plus the git commit.

For example, version today would be:

2.1.0-dev+gitXXXXXX

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>