Introduce the `display_name` option
From now on, users will be able to introduce a "display name". This display
name will be the one shown on the website. The main goal of this feature is to
allow users to see names on the website that are more familiar to them (e.g.
users coming from LDAP but that have an "ugly" username there).

Notice that this "display name" is only used on the web site, not in the
`docker login` procedure. Because of this, and because it might confuse people,
this configuration option is disabled by default.

Fixes #699

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>