add option to disable change of visibility
There is a new option 'user_change_visibility' which allows users to
change the visibility of their personal namespace. If this is option is
disabled, only admins can perform this change. This option is enabled by
default and therefore doesn't change the current behavior.

This commit is part of issue #676.

Signed-off-by: Thomas Hipp <thipp@suse.de>