namespace: convert to downcase first on make_valid
One usual pain point that can be solved in `make_valid` is that lots of times
the problem is with the case. Since upper case is not accepted in namespace
names, then it will be converted weirdly. Because of this, from now on the
first thing that `make_valid` will do if the name doesn't match right away, is
to convert the name to downcase. With this commit then, something like `Miquel`
will be converted to `miquel`, instead of `_iquel`, which makes more sense.

See #965

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>