Added the registry.catalog_page option
With this option we will be able to workaround docker/distribution#1854.
Moreover, with this we will allow administrators to further tweak how Portus
interacts with the registry.

Moreover, because of this new addition, I felt like groupping registry-related
configurable options would be a good idea, so now all these options are under
the new `registry` section. We are still providing a compatibility layer in the
Portus::Migrate module.

Fixes #1001

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>