Fixed some bugs related to activities
Besides the mentioned issues, this commit also fixes some activities when
deleting users. This commit also fixes most of the issues in #899, but
since this wasn't the goal of this commit, fixing this issue has been
left as something to be done in the near future (I'd still need to
properly review it).

Fixes #1104
Fixes #1106
See #899

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>