Moved from Coffeescript to Javascript
With Ecmascript 6, the Javascript language is not as terrible as it used
to be (well, it still is a mess), so languages like Coffeescript are no
longer needed. Getting rid of this dependency means that it should be
easier to deploy (e.g. no execjs gem). Thus, it will give us thinner
deployments with the exact same functionality.

That being said, I've moved to Ecmascript 5 instead of 6, because not
all browsers support it. I could've integrated Babel into the
development, but it all seems immature in Rails 4.2 (e.g. no sprocket
support for this version to do the conversion automatically in
development). It may be possible, but I didn't want to get into too much
trouble for now.

I've done the move with the decaffeinate tool, and then translated it
into ecmascript 5 with Babel (I've left the configuration of Babel on
the package.json file as well). I've also added ESLint, which will be
run on Travis CI. The style has been borrowed from Airbnb and Gitlab,
but I don't have any strong opinions on the specific style really.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>