Return empty tags on unsuccessful catalogs
Whenever we try to fetch the tags of a given repository, set the "tags"
value to nil for repositories for which the request was unsuccessful.
This is already handled by the CatalogJob as a way to skip it. This way,
we avoid removing repos when a 401 response has been returned when
accessing the catalog.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>