crono: don't let crono run the migrations
In production it might happen that the crono process is started before
the portus process (e.g. a containerized setup with docker-compose or
Kubernetes). In such a case, we have to be sure that the process running
the migration is the main Portus one, because otherwise the `init`
script from portus won't be able to run successfully, and thus it won't
ever run `rake db:seed` on initialization.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>