production: make the log level configurable
With this commit administrators will be able to set the log level of
their Portus instance through the `PORTUS_LOG_LEVEL`. This will be
possible in production and staging, and it's meant to be used in
exceptional cases (e.g. something is really wrong with the `crono`
process and the administrator wants to restart it with `debug` mode for
a while).

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>