js: added table sortable mixin; namespaces using it
So far the column sorting was possible but user wasn't able
to share the state of it to someone through copying and pasting url.

Now every sorting action the user makes on the namespaces table,
for example, will create a new url state that can be visited and
navigated back and forth.

Fixes #1373