js: added table paginated mixin; namespaces using it
So far the pagination was possible but user wasn't able
to share the state of it with someone by copying and pasting url.

Now every pagination action the user makes on the namespaces table,
for example, will create a new url state that can be visited and
navigated back and forth.

Fixes #1390