js: added inline validation for namespace's new form
The validation so far was fully done on the server side and
the feedback was through the float alert that disappears in 5 seconds.
Not the best experience for the users.

Now, for each field (name and team), we have an inline feedback validation.
User will only see a validation message on the float alert in unexpected
situations.

Fixes #1376