js: added some registry validation and status
The registry validation was done completely on the server side. With
this, we bring some validation to the client side to antecipate and improve
the user's experience.

Also adds a hint in the registries#index pages to help the user know if the
registry that was set is reachable or not. The reachable status is cached for
30 minutes unless the user updates the registry.

Fixes #1355