Added PostgreSQL support
The database configuration has been extended to also support PostgreSQL
as the database. This is relevant if you are using CoreOS Clair as the
security scanner and you don't want to have multiple databases.

In theory with this implementation we could have support for other RDBMs
as well, but that would mean to add gems for each support and call
`bundle`. To avoid these kinds of scenarios, we stick with just two options:
with MariaDB (the currently supported option) and PostgreSQL.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>