Properly check SSL usage requirements
In a couple of places (around the API) there was a check for the
PORTUS_CHECK_SSL_USAGE_ENABLED environment variable. Doing this is wrong
because it skips the fact that this is the default behavior as defined
in the main config.yml file. So, the `APP_CONFIG.enabled?` method should
be used instead.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>