Added custom validator for the uniqueness of tags
Even if the case sensitivity is set to true when you add a uniqueness
constraint in Ruby on Rails, it happens that in MariaDB/MySQL this is
not correctly handled when this constraint is coupled with another
column. In our case, the uniqueness constraint was between tag name and
repository. In this case, in MariaDB the evaluated key for the
uniqueness index was case insensitive (while in PostgreSQL everything
went as expected). This means that for two tags named "tag" and "TAG"
and inside of the repository "4", the evaluated keys were "tag-4" and
"TAG-4", and in MariaDB they were considered equal, hence it failed.

I've tried multiple solutions, but they all failed:

- Explicitly setting the case_sensitive property on the Rails' side.
- Changing the collation for the table, column, and combinations. This
  was successful locally but not on the CI server. Moreover, it broke
  the schema:load task.

The only solution left to us was to drop the `uniqueness` validator (and
the index with it) and write a custom one. This will certainly be slightly
inefficient, but it looks like we have no other option. Just to be sure,
I've added tests for this whole constraint mess, and for the other
constraints.

In the future we might evaluate removing MySQL support, since it's not
the first time that it has some downsides for the overall source code.
This is something to consider on the next major versions (where
breakages and manual migrations are to be expected).

Fixes #1262

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>