api: fixed crash on null author of a tag
It's not guaranteed from the DB that the author cannot be null.
Moreover, it happened to me that the author of a tag was nil, and so
Portus crashed when requesting this info through the API.

This commit simply sets the author info to null if the author is also
null.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>