api: route other tasks to a proper 404
In order to do this, I had to move the Swagger documentation because it
was mounted by Rails and not Grape, and there was a conflict if it was
under the /api path.

The `route :any, "*path"` solution comes from the Grape documentation,
and it looks like it works fine.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>