Fixed blocking issues with security scanning
Unless you provided enough Puma workers, there was a problem with HTTP
requests getting blocked. This is a similar issue as pointed out in #1353
with registry events.

The solution has been to move security scanning in the background, in a
different process. This is similar to #1353, but more polished (and if
this proves to really work we might move #1353 to this background runner
as well). Now the Tag model has two more columns: `vulnerabilities` and
`scanned`. The former contains the cached result of the latest security
check. The latter is an integer that can have three different values: 0
(scanning not performed), 1 (working on it) and 2 (scanning done). This
has also been exposed on the API, so the client side can be updated.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>