rubocop: bring it closer to upstream
After some discussions internally, we decided that it was good to be as
close to the default upstream rubocop configuration. For this reasons,
I have trimmed down as much as possible our own modifications for the
rubocop file, and hence we had to re-adapt a lot of code.

There are some things that have been left out:

- Some extra cops like rspec-rubocop.
- Further trimming down some configuration options that needed some
  heavy liftings...

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>